requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
chromadb==0.4.15
tiktoken==0.5.1
python-dotenv==1.0.0
pyyaml==6.0.1
